requirements.txt
tensorflow~=2.2.0
numpy~=1.19.0
Pillow~=8.0.1
matplotlib~=3.3.2